requirements.txt
asgiref==3.11.0
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi @ file:///private/tmp/python-certifi-20231008-5561-13s00fm/certifi-2023.7.22
cffi @ file:///private/tmp/cffi-20231014-5567-1hyue8f/cffi-1.16.0
charset-normalizer==3.3.2
distlib==0.3.6
Django==5.2.9
filelock==3.12.2
h11==0.14.0
idna==3.6
numpy==1.26.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.1.3
platformdirs==3.9.0
psutil @ file:///private/tmp/python-psutil-20231015-5378-1ng7p6p/psutil-5.9.6
pycparser @ file:///private/tmp/pycparser-20231013-5787-1anpge2/pycparser-2.21
Pygments @ file:///private/tmp/pygments-20231008-5534-zuvwpt/Pygments-2.16.1
pyserial==3.5
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
PySimpleGUI==4.60.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML @ file:///private/tmp/pyyaml-20231006-5676-18tnnkv/PyYAML-6.0.1
requests==2.31.0
selenium==4.15.2
shiboken6==6.5.1.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.4
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.0
vue==0.0.1
webdriver-manager==4.0.1
wsproto==1.2.0
